# src/main/resources/api/openapi.yaml
openapi: 3.0.2
info:
  title: FG-AI4H-OCI Platform
  description: >
    'FG-AI4H Open Code Initiative Platform'
  termsOfService: https://github.com/FG-AI4H/annotation-tool/blob/master/LICENSE
  contact:
    email: ml@mllab.ai
  license:
    name: BSD 3-Clause License
    url: https://github.com/FG-AI4H/annotation-tool/blob/master/LICENSE
  version: 2023-02-20
servers:
  - url: https://annotation.ai4h.net/api/v1
    description: Production server
  - url: http://localhost:8080/api/v1
    description: Development server
...
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DatasetDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        storage_location:
          type: string
        linked:
          type: boolean
        catalog_location:
          type: string
        catalog_auth_type:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
        data_catalog_id:
          type: string
          format: uuid
        metadata:
          $ref: '#/components/schemas/DatasetMetadataDto'
        visibility:
          $ref: '#/components/schemas/DatasetVisibility'
        request_status:
          $ref: '#/components/schemas/DatasetCatalogRequestStatusDto'
    DatasetMetadataDto:
      type: object
      properties: 
        id:
          type: string
          format: uuid
        data_collection_authors_agency:
          type: string
        data_acceptance_standards_compliance:
          type: string
        data_acquisition_sensing_device_type:
          type: string
        data_acquisition_sensing_modality:
          type: string
        data_annotation_process_tool:
          type: string
        data_assumptions_constraints_dependencies:
          type: string
        data_bias_and_variance_minimization:
          type: string
        data_collection_funding_agency:
          type: string
        data_collection_period:
          type: string
        data_collection_place:
          type: string
        data_dimension:
          type: string
        data_exclusion_criteria:
          type: string
        data_owner_id:
          type: string
          format: uuid
        data_owner_name:
          type: string
        data_preprocessing_techniques:
          type: string
        data_privacy_de_identification_protocol:
          type: string
        data_registry_url:
          type: string
        data_resolution_precision:
          type: string
        data_safety_security_protocol:
          type: string
        data_sample_size:
          type: string
        data_sampling_rate:
          type: string
        data_source:
          type: string
        data_type:
          type: string
        data_update_version:
          type: string
        train_tuning_eval_dataset_partitioning_ratio:
          type: string
        version:
          type: string
    DatasetVisibility:
      type: string
      enum:
        - PRIVATE
        - PUBLIC
    DatasetRoleDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        created_at:
          type: string
        updated_at:
          type: string
        dataset:
          type: string
          format: uuid
        user:
          type: string
          format: uuid
        username:
          type: string
        user_role:
          $ref: '#/components/schemas/UserRole'
    DatasetCatalogRequestStatusDto:
      type: string
      enum:
        - PENDING
        - APPROVED
        - REJECTED
        - DEPRECATED
    UserRole:
      type: string
      enum:
        - OWNER
        - CONTRIBUTOR
        - READER
    CampaignDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/CampaignStatus'
        annotation_kind:
          type: string
        annotation_tool:
          type: string
        pre_annotation_tool:
          type: string
        pre_annotation_model:
          type: string
        annotation_method:
          type: string
        annotation_instructions:
          type: string
        quality_assurance:
          type: string
        is_instance_label:
          type: boolean
          default: false
        min_annotation:
          type: integer
        class_label:
          type: array
          items:
            $ref: '#/components/schemas/ClassLabelDto'
        datasets:
          type: array
          items:
            type: string
            format: uuid
        annotators:
          items:
            type: string
            format: uuid
        reviewers:
          items:
            type: string
            format: uuid
        supervisors:
          items:
            type: string
            format: uuid
    CampaignStatus:
      type: string
      enum:
        - DRAFT
        - READY
        - RUNNING
        - COMPLETED
        - ARCHIVED
    ClassLabelDto:
      type: object
      description: A class label
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
...
    AnnotationTaskDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        kind:
          type: string
        title:
          type: string
        description:
          type: string
...
    AnnotationToolDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        editor:
          type: string
        annotation_tasks:
          type: array
          items:
            $ref: '#/components/schemas/AnnotationTaskDto'
...
paths:
...
  /campaigns/{campaign_id}:
    get:
      tags:
        - Campaign
      summary: Returns a campaign
      description: Returns the campaign of given campaignId
      operationId: getCampaignById
      parameters:
        - name: campaign_id
          in: path
          description: Campaign Identifier
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignDto'
        '404':
          description: Given campaignId ID doesn't exist
          content: { }
    put:
      tags:
        - Campaign
      summary: Updates a new campaign for the given campaign request
      description: Updates a new campaign for the given campaign request.
      operationId: updateCampaign
      parameters:
        - name: campaign_id
          in: path
          description: Campaign Identifier
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Campaign Request object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampaignDto'
      responses:
        '204':
          description: Campaign updates successfully
        '500':
          description: Error occured
          content: { }
    delete:
      tags:
        - Campaign
      summary: Delete a campaign
      description: Delete the campaign of given campaignId
      operationId: deleteCampaignById
      parameters:
        - name: campaign_id
          in: path
          description: Campaign Identifier
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: successful operation
        '404':
          description: Given campaignId ID doesn't exist
          content: { }
...
  /datasets:
    get:
      tags:
        - Dataset
      summary: List all the datasets
      description: List all the datasets
      operationId: getAllDatasets
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DatasetDto'
        '404':
          description: Error occured
          content: { }
    post:
      tags:
        - Dataset
      summary: Creates a new dataset for the given dataset request
      description: Creates a new dataset for the given dataset request.
      operationId: addDataset
      requestBody:
        description: New dataset Request object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatasetDto'
      responses:
        '201':
          description: Dataset added successfully
  /datasets/catalog:
    get:
      tags:
        - Dataset
      summary: List all the datasets from all catalog
      description: List all the datasets from all catalog
      operationId: getCatalogDatasets
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DatasetDto'
        '404':
          description: Error occured
          content: { }
  /datasets/catalog/request:
    post:
      tags:
        - Dataset
      summary: Creates request to link dataset to dataset owner
      description: Creates request to link dataset to dataset owner
      operationId: requestAccessToCatalogDataset
      requestBody:
        description: Dataset catalog Request object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatasetDto'
      responses:
        '201':
          description: Request created successfully
  /datasets/{dataset_id}:
    get:
      tags:
        - Dataset
      summary: Returns a dataset
      description: Returns the dataset of given datasetId
      operationId: getDatasetById
      parameters:
        - name: dataset_id
          in: path
          description: Dataset Identifier
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetDto'
        '404':
          description: Given Dataset ID doesn't exist
          content: { }
    put:
      tags:
        - Dataset
      summary: Update the dataset
      description: Update the given dataset
      operationId: updateDatatset
      parameters:
        - name: dataset_id
          in: path
          description: Dataset Identifier
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: New Dataset object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatasetDto'
      responses:
        '204':
          description: Datset updated successfully
        '500':
          description: Given dataset ID doesn't exist
          content: { }
    delete:
      tags:
        - Dataset
      summary: Delete a dataset
      description: Delete the dataset of given datasetId
      operationId: deleteDatasetById
      parameters:
        - name: dataset_id
          in: path
          description: Dataset Identifier
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: successful operation
        '404':
          description: Given Dataset ID doesn't exist
          content: { }
  /datasets/{dataset_id}/permissions:
    get:
      tags:
        - Dataset
      summary: Returns a dataset permissions
      description: Returns the dataset permissions of given datasetId
      operationId: getDatasetPermissionsById
      parameters:
        - name: dataset_id
          in: path
          description: Dataset Identifier
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DatasetRoleDto'
        '404':
          description: Given Dataset ID doesn't exist
          content: { }
...
  /annotation_tools:
    get:
      tags:
        - Tool
      summary: Get annotation tools
      description: Returns the list of annotation tools
      operationId: getAnotationTools
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AnnotationToolDto'
        '404':
          description: Error
          content: { }
    post:
      tags:
        - Tool
      summary: Creates a new annotation tool for the given tool request
      description: Creates a new annotation tool for the given tool request
      operationId: addAnnotationTool
      requestBody:
        description: New Tool Request object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnnotationToolDto'
      responses:
        '201':
          description: Tool added successfully
  /annotation_tools/{annotation_tool_id}:
    get:
      tags:
        - Tool
      summary: Get annotation tool
      description: Returns annotation tool corresponding to the uuid
      operationId: getAnnotationToolById
      parameters:
        - name: annotation_tool_id
          in: path
          description: Annotation Tool Identifier
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnotationToolDto'
        '404':
          description: Error
          content: { }
...